+++ saleor/static/dashboard/js/components/app/category/edit.js
@@ -1,22 +1,8 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import gql from 'graphql-tag';
 import { graphql } from 'react-apollo';
 import CategoryPropertiesForm from './categoryPropertiesForm';
-
-const query = gql`
-  query CategoryDetails($pk: Int!) {
-    category(pk: $pk) {
-      id
-      pk
-      name
-      description
-      parent {
-        pk
-      }
-    }
-  }
-`;
+import { CategoryDetails as query } from './queries';
 
 export default graphql(query, {
   options: (props) => ({
