+++ saleor/static/dashboard/js/components/app/category/queries.js
@@ -1,0 +1,34 @@
+import gql from 'graphql-tag';
+
+const CategoryChildren = gql`
+  query CategoryChildren ($pk: Int) {
+    categories(parent: $pk) {
+      edges {
+        node {
+          id
+          pk
+          name
+          description
+        }
+      }
+    }
+  }
+`;
+const CategoryDetails = gql`
+  query CategoryDetails($pk: Int!) {
+    category(pk: $pk) {
+      id
+      pk
+      name
+      description
+      parent {
+        pk
+      }
+    }
+  }
+`;
+
+export {
+  CategoryChildren,
+  CategoryDetails
+};
